Extract shared time utilities and clean up duplicated code
- Created lib/timeUtils.ts with timeToMinutes, windowDurationMinutes,
  and formatDurationLabel
- Replaced 3 inline duration calculations (PollPageClient, create-poll,
  DayTimeWindowsInput) with windowDurationMinutes()
- Replaced 3 inline duration label formatters with formatDurationLabel()
- Removed unused deferValidation prop from MinMaxCounter
- Simplified double-negation in toggle handler
- Removed unnecessary prop alias in TimeGridModal

Net: -27 lines
